timeout: fixup

+++ userbot/modules/system_stats.py
@@ -24,7 +24,7 @@ import psutil
 from git import Repo
 from telethon import __version__, version
 
-from userbot import ALIVE_LOGO, ALIVE_NAME, CMD_HELP, USERBOT_VERSION, StartTime, bot
+from userbot import ALIVE_LOGO, ALIVE_NAME, CMD_HELP, TIMEOUT, USERBOT_VERSION, StartTime, bot
 from userbot.events import register
 
 # ================= CONSTANT =================
@@ -234,16 +234,18 @@ async def amireallyalive(alive):
     uptime = await get_readable_time((time.time() - StartTime))
     output = (
         f"`===============================`\n"
-        f"**Fizilion is Up and Running.... **\n"
-        f"`=============================== `\n\n"
-        f"[**OS Info**]:\n\n"
+        f"**FIZILION IS UP AND RUNNING...**\n"
+        f"`=============================== `\n"
+        f"**[OS Info]:**\n"
         f"•`Platform Type   : {os.name}`\n"
-        f"•`Distro          : {distro.name(pretty=False)} {distro.version(pretty=False, best=False)}`\n\n"
-        f"[**PYPI Module Versions**]:\n\n"
+        f"•`Distro          : {distro.name(pretty=False)} {distro.version(pretty=False, best=False)}`\n"
+        f"`===============================`\n"
+        f"**[PYPI Module Versions]:**\n"
         f"•`Python         : v{python_version()} `\n"   
         f"•`Telethon       : v{version.__version__} `\n"
-        f"•`PIP            : v{pip.__version__} `\n\n"
-        f"[**MISC Info**]:\n\n"
+        f"•`PIP            : v{pip.__version__} `\n"
+        f"`===============================`\n"
+        f"**[MISC Info]:**\n"
         f"•`User           : {DEFAULTUSER} `\n"
         f"•`Running on     : {repo.active_branch.name} `\n"
         f"•`Loaded modules : {len(modules)} `\n"
@@ -257,8 +259,7 @@ async def amireallyalive(alive):
             logo = ALIVE_LOGO
             msg = await bot.send_file(alive.chat_id, logo, caption=output, del_in=10)
             await alive.delete()
-            await sleep(15)
-            await msg.delete()
+            await sleep(30)
         except BaseException:
             await alive.edit(
                 output + "\n\n *`The provided logo is invalid."
@@ -266,9 +267,11 @@ async def amireallyalive(alive):
             )
     else:
         msg=await alive.edit(output)
-        await sleep(15)
+        await sleep(30)
+        
+    if TIMEOUT:
         await msg.delete()
-
+        
 
 @register(outgoing=True, pattern="^.aliveu")
 async def amireallyaliveuser(username):
